Refactor crowd sim, remove common package
