--- rmf_building_sim_common/README.md
@@ -1,7 +1,0 @@
-# rmf\_building\_sim\_common
-
-This package provides a common library for the Ignition and Gazebo simulator plugins for simulated building infrastructure.
-
-## Quality Declaration
-
-This package claims to be in the **Quality Level 4** category, see the [Quality Declaration](./QUALITY_DECLARATION.md) for more details.
