--- rmf_building_sim_common/package.xml
@@ -1,31 +1,0 @@
-<package format="3">
-  <name>rmf_building_sim_common</name>
-  <version>2.2.2</version>
-  <description>
-    Common utility functions for Gazebo-classic and Gazebo building plugins
-  </description>
-  <author>Morgan Quigley</author>
-  <author>Michael Grey</author>
-  <author>Roselle Carmen</author>
-  <author>Aaron Chong</author>
-  <author>Brandon Ong</author>
-  <author>Kevin Ma</author>
-  <author>Rushyendra Maganty</author>
-  <author>Yadunund</author>
-  <maintainer email="morgan@openrobotics.org">Morgan Quigley</maintainer>
-  <maintainer email="marco@openrobotics.org">Marco A. Gutiérrez</maintainer>
-  <license>Apache 2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>rclcpp</depend>
-  <depend>rmf_door_msgs</depend>
-  <depend>rmf_lift_msgs</depend>
-  <depend>rmf_building_map_msgs</depend>
-  <depend>menge_vendor</depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-
-</package>
